added models subpackage
